requirements.txt
fastapi
uvicorn
pydantic
pylance
sqlalchemy
psycopg2-binary
PyMySQL
cryptography
python-dotenv
AzureOpenAI
openai
azure-search-documents
python-multipart